Disable name_list shortcut for pending names.
For names that have a pending wallet transaction, always do the "full
check" in name_list and don't try to shortcut them when all outputs are
spent.  Even the pending tx counts as "spending" of the outputs, while
it doesn't yet show up later for name_list, so that the name effectively
goes missing.