Make address reuse for movements controllable.
Add a new "-noaddressreuse" command-line parameter that disables the
reuse of player addresses for movements.  By default, addresses will be
reused to prevent running out of backed up keys.  This patch also
changes the behaviour to send not only the player itself but also the
change to a reused address.

This fixes https://github.com/chronokings/huntercoin/issues/67.