Re-introduce GetTxOfNameAtHeight with sanity check.
Move the sanity check for game state height from CreateGameTransactions
(where it fails for reorgs) to GetTxOfNameAtHeight.  In constrast to the
previous implementation of GetTxOfNameAtHeight, this one can only
retrieve the latest entry but errors out if the height requested is too
early.  This should not happen with the current implementation of
CreateGameTransactions.