Add sanity check for coin amount after game step.
Add a new check when performing a game step:  Compute the expected money
in and out of the game world, and verify that the total coins in the
world before and after the step match the expectations.  This may help
to prevent exploits that could be used to create more coins.  If such a
situation happens, the corresponding block is not accepted.